Initial commit
Plugin which has a settings page. That page has one button, which
allows the user to link or unlink their site to LinkedIn. Linking
means that the user will go through an authorization process,
which results in an auth code specific to a LinkedIn account
being saved locally. Unlinking removes this saved entry.